linkify subject + open xmpp directly w/o going through start conv activity
