+++ src/main/res/layout/activity_muc_details.xml
@@ -73,6 +73,7 @@
                                         android:layout_width="wrap_content"
                                         android:layout_height="wrap_content"
                                         android:layout_alignParentStart="true"
+                                        android:autoLink="web"
                                         android:layout_below="@+id/muc_title"
                                         android:layout_toStartOf="@+id/edit_muc_name_button"
                                         android:textAppearance="@style/TextAppearance.Conversations.Subhead"/>
